requirements.txt
auth0-python==3.9.1
autopep8==1.4.4
flake8==3.7.9